#561 Extract hidden inner class `CommandElementContext`
